Drive
Made it possible to drive the robot with no real problems

+++ Functions.c
@@ -1,4 +1,4 @@
-int PID(int current, int desired, int pGain, int iGain, int dGain, int safety){
+int PID(int current, int desired, float pGain, float iGain, float dGain, float safety){
 	int pCurrent;
 	int iCurrent;
 	int dCurrent;
